# swagger-codegen/go/model_pet.go
/*
 * Swagger Example
 *
 * Swagger Example
 *
 * API version: 1.0.6
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Pet struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name"`
}